Split up Auto Splitter API into multiple files
This splits up the Auto Splitter API into multiple files. The API has
significantly grown in size and it's hard to find anything in the
single file. This splits it up into multiple files, so that it's easier
to find things.
